Update FABulous folder structure for better API update
