+++ .github/workflows/fabric_gen.yml
@@ -1,49 +1,48 @@
 name: Test fabric_gen.py
 
 on:
-    push:
-        branches: [master]
-    pull_request:
-        branches: [master]
+  push:
+    branches: [master]
+  pull_request:
+    branches: [master]
 
 jobs:
-    build_and_run:
-        runs-on: ubuntu-latest
+  build_and_run:
+    runs-on: ubuntu-latest
 
-        steps:
-            - uses: actions/checkout@v4
-              with:
-                  submodules: recursive
-            - name: Set up Python 3.9
-              uses: actions/setup-python@v4
-              with:
-                  python-version: 3.9
-            - name: Set up OSS CAD suite
-              uses: YosysHQ/setup-oss-cad-suite@v2
-            - name: Install dependencies
-              run: |
-                  python3 -m pip install --upgrade pip
-                  pip3 install flake8 pytest
-                  if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-                  if [ -f requirements.txt ]; then python3.9 -m pip install -r requirements.txt; fi
-                  sudo apt update
-                  sudo apt install python-tk -y
-            - name: Lint with flake8
-              run: |
-                  # stop the build if there are Python syntax errors or undefined names
-                  flake8 FABulous/**/*.py --count --select=E9,F63,F7,F82 --show-source --statistics
-                  # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-                  flake8 FABulous/**/*.py --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-            - name: Install FABulous
-              run: |
-                  pip3 install -e .
-            - name: Run fabric generator flow
-              run: |
-                  FABulous -c demo
-                  FABulous demo --script ./demo/FABulous.tcl
-            - name: Run simulation smoketest
-              run: |
-                  cd ./demo/Test
-                  ./build_test_design.sh
-                  ./run_simulation.sh
-
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: recursive
+      - name: Set up Python 3.10
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.10
+      - name: Set up OSS CAD suite
+        uses: YosysHQ/setup-oss-cad-suite@v2
+      - name: Install dependencies
+        run: |
+          python3 -m pip install --upgrade pip
+          pip3 install flake8 pytest
+          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+          if [ -f requirements.txt ]; then python3.10 -m pip install -r requirements.txt; fi
+          sudo apt update
+          sudo apt install python-tk -y
+      - name: Lint with flake8
+        run: |
+          # stop the build if there are Python syntax errors or undefined names
+          flake8 FABulous/**/*.py --count --select=E9,F63,F7,F82 --show-source --statistics
+          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+          flake8 FABulous/**/*.py --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+      - name: Install FABulous
+        run: |
+          pip3 install -e .
+      - name: Run fabric generator flow
+        run: |
+          FABulous -c demo
+          FABulous demo --script ./demo/FABulous.tcl
+      - name: Run simulation smoketest
+        run: |
+          cd ./demo/Test
+          ./build_test_design.sh
+          ./run_simulation.sh
