+++ FABulous/fabric_definition/define.py
@@ -1,0 +1,34 @@
+from enum import Enum
+
+
+class IO(Enum):
+    INPUT = "INPUT"
+    OUTPUT = "OUTPUT"
+    INOUT = "INOUT"
+    NULL = "NULL"
+
+
+class Direction(Enum):
+    NORTH = "NORTH"
+    SOUTH = "SOUTH"
+    EAST = "EAST"
+    WEST = "WEST"
+    JUMP = "JUMP"
+
+
+class Side(Enum):
+    NORTH = "NORTH"
+    SOUTH = "SOUTH"
+    EAST = "EAST"
+    WEST = "WEST"
+    ANY = "ANY"
+
+
+class MultiplexerStyle(Enum):
+    CUSTOM = "CUSTOM"
+    GENERIC = "GENERIC"
+
+
+class ConfigBitMode(Enum):
+    FRAME_BASED = "FRAME_BASED"
+    FLIPFLOP_CHAIN = "FLIPFLOP_CHAIN"
