+++ FABulous/fabric_generator/code_generation_Verilog.py
@@ -3,7 +3,7 @@ import re
 from typing import Literal
 
 from FABulous.fabric_generator.code_generator import codeGenerator
-from FABulous.fabric_generator.fabric import IO, Bel, ConfigBitMode, Tile
+from FABulous.fabric_definition.Fabric import IO, Bel, ConfigBitMode, Tile
 
 
 class VerilogWriter(codeGenerator):
