+++ FABulous/fabric_generator/file_parser.py
@@ -2,21 +2,24 @@ import csv
 import os
 import re
 from copy import deepcopy
-from typing import Dict, List, Literal, Tuple, Union, overload
 
-from FABulous.fabric_generator.fabric import (
+from FABulous.fabric_definition.Bel import Bel
+from FABulous.fabric_definition.Port import Port
+from FABulous.fabric_definition.Wire import Wire
+from FABulous.fabric_definition.Tile import Tile
+from FABulous.fabric_definition.SuperTile import SuperTile
+from FABulous.fabric_definition.Fabric import Fabric
+from FABulous.fabric_definition.ConfigMem import ConfigMem
+from FABulous.fabric_generator.utilities import parseList, parseMatrix
+from typing import Literal
+from FABulous.fabric_definition.define import (
     IO,
-    Bel,
-    ConfigBitMode,
-    ConfigMem,
     Direction,
-    Fabric,
-    MultiplexerStyle,
-    Port,
     Side,
-    SuperTile,
-    Tile,
+    ConfigBitMode,
+    MultiplexerStyle,
 )
+
 
 # from fabric import Fabric, Port, Bel, Tile, SuperTile, ConfigMem
 # from fabric import IO, Direction, Side, MultiplexerStyle, ConfigBitMode
@@ -82,12 +85,12 @@ def parseFabricCSV(fileName: str) -> Fabric:
     # Lists for tiles
     tileTypes = []
     tileDefs = []
-    commonWirePair: List[Tuple[str, str]] = []
+    commonWirePair: list[tuple[str, str]] = []
 
     fabricTiles = []
     tileDic = {}
 
-    # List for supertiles
+    # list for supertiles
     superTileDic = {}
 
     # For backwards compatibility parse tiles in fabric config
@@ -220,7 +223,7 @@ def parseFabricCSV(fileName: str) -> Fabric:
     )
 
 
-def parseTiles(fileName: str) -> Tuple[List[Tile], List[Tuple[str, str]]]:
+def parseTiles(fileName: str) -> tuple[list[Tile], list[tuple[str, str]]]:
     """
     Parses a csv tile configuration file and returns all tile objects.
 
@@ -228,7 +231,7 @@ def parseTiles(fileName: str) -> Tuple[List[Tile], List[Tuple[str, str]]]:
         fileName (str): the path to the csv file.
 
     Returns:
-        ([Tile], [(str, str)] : Tuple of tile objects and common wire pairs.
+        ([Tile], [(str, str)] : tuple of tile objects and common wire pairs.
     """
 
     print(f"Reading tile configuration: {fileName}")
@@ -254,13 +257,13 @@ def parseTiles(fileName: str) -> Tuple[List[Tile], List[Tuple[str, str]]]:
     for t in tilesData:
         t = t.split("\n")
         tileName = t[0].split(",")[1]
-        ports: List[Port] = []
-        bels: List[Bel] = []
+        ports: list[Port] = []
+        bels: list[Bel] = []
         matrixDir = ""
         withUserCLK = False
         configBit = 0
         for item in t:
-            temp: List[str] = item.split(",")
+            temp: list[str] = item.split(",")
             if not temp or temp[0] == "":
                 continue
             if temp[0] in ["NORTH", "SOUTH", "EAST", "WEST"]:
@@ -384,7 +387,7 @@ def parseTiles(fileName: str) -> Tuple[List[Tile], List[Tuple[str, str]]]:
     return (new_tiles, commonWirePair)
 
 
-def parseSupertiles(fileName: str, tileDic: Dict[str, Tile]) -> List[SuperTile]:
+def parseSupertiles(fileName: str, tileDic: dict[str, Tile]) -> list[SuperTile]:
     """
     Parses a csv supertile configuration file and returns all SuperTile objects.
 
@@ -393,7 +396,7 @@ def parseSupertiles(fileName: str, tileDic: Dict[str, Tile]) -> List[SuperTile]:
         tileDic ({str: Tile}): dict of tiles.
 
     Returns:
-        [SuperTile]: List of supertile objects.
+        [SuperTile]: list of supertile objects.
     """
 
     print(f"Reading supertile configuration: {fileName}")
@@ -473,118 +476,14 @@ def parseSupertiles(fileName: str, tileDic: Dict[str, Tile]) -> List[SuperTile]:
     return new_supertiles
 
 
-@overload
-def parseList(
-    fileName: str, collect: Literal["pair"] = "pair"
-) -> List[Tuple[str, str]]:
-    pass
-
-
-@overload
-def parseList(
-    fileName: str, collect: Literal["source", "sink"]
-) -> Dict[str, List[str]]:
-    pass
-
-
-def parseList(
-    fileName: str, collect: Literal["pair", "source", "sink"] = "pair"
-) -> Union[List[Tuple[str, str]], Dict[str, List[str]]]:
-    """
-    parse a list file and expand the list file information into a list of tuples.
-
-    Args:
-        fileName (str): ""
-        collect (Literal[&quot;&quot;, &quot;source&quot;, &quot;sink&quot;], optional): Collect value by source, sink or just as pair. Defaults to "pair".
-
-    Raises:
-        ValueError: The file does not exist.
-        ValueError: Invalid format in the list file.
-
-    Returns:
-        Union[List[Tuple[str, str]], Dict[str, List[str]]]: Return either a list of connection pair or a dictionary of lists which is collected by the specified option, source or sink.
-    """
-
-    if not os.path.exists(fileName):
-        raise ValueError(f"The file {fileName} does not exist.")
-
-    resultList = []
-    with open(fileName, "r") as f:
-        file = f.read()
-        file = re.sub(r"#.*", "", file)
-    file = file.split("\n")
-    for i, line in enumerate(file):
-        line = line.replace(" ", "").replace("\t", "").split(",")
-        line = [i for i in line if i != ""]
-        if not line:
-            continue
-        if len(line) != 2:
-            print(line)
-            raise ValueError(f"Invalid list formatting in file: {fileName} at line {i}")
-        left, right = line[0], line[1]
-
-        leftList = []
-        rightList = []
-        _expandListPorts(left, leftList)
-        _expandListPorts(right, rightList)
-        resultList += list(zip(leftList, rightList))
-
-    result = list(dict.fromkeys(resultList))
-    resultDic = {}
-    if collect == "source":
-        for k, v in result:
-            if k not in resultDic:
-                resultDic[k] = []
-            resultDic[k].append(v)
-        return resultDic
-
-    if collect == "sink":
-        for k, v in result:
-            for i in v:
-                if i not in resultDic:
-                    resultDic[i] = []
-                resultDic[i].append(k)
-        return resultDic
-
-    return result
-
-
-def _expandListPorts(port, PortList):
-    """
-    expand the .list file entry into list of tuple.
-    """
-    # a leading '[' tells us that we have to expand the list
-    if "[" in port:
-        if "]" not in port:
-            raise ValueError(
-                "\nError in function ExpandListPorts: cannot find closing ]\n"
-            )
-        # port.find gives us the first occurrence index in a string
-        left_index = port.find("[")
-        right_index = port.find("]")
-        before_left_index = port[0:left_index]
-        # right_index is the position of the ']' so we need everything after that
-        after_right_index = port[(right_index + 1) :]
-        ExpandList = []
-        ExpandList = re.split(r"\|", port[left_index + 1 : right_index])
-        for entry in ExpandList:
-            ExpandListItem = before_left_index + entry + after_right_index
-            _expandListPorts(ExpandListItem, PortList)
-
-    else:
-        # print('DEBUG: else, just:',port)
-        PortList.append(port)
-    return
-
-
-def parseFileVHDL(filename: str, belPrefix: str = "") -> Tuple[
-    List[Tuple[str, IO]],
-    List[Tuple[str, IO]],
-    List[Tuple[str, IO]],
-    List[Tuple[str, IO]],
+def parseFileVHDL(filename: str, belPrefix: str = "") -> tuple[
+    list[tuple[str, IO]],
+    list[tuple[str, IO]],
+    list[tuple[str, IO]],
+    list[tuple[str, IO]],
     int,
     bool,
-    Dict[str, int],
+    dict[str, int],
 ]:
     """
     Parse a VHDL bel file and return all the related information of the bel. The tuple returned for relating to ports will
@@ -602,14 +501,14 @@ def parseFileVHDL(filename: str, belPrefix: str = "") -> Tuple[
         ValueError: Cannot find the port section in the file which defines the bel ports.
 
     Returns:
-        Tuple[List[Tuple[str, IO]], List[Tuple[str, IO]], List[Tuple[str, IO]], List[Tuple[str, IO]], int, bool, Dict[str, int]]:
+        tuple[list[tuple[str, IO]], list[tuple[str, IO]], list[tuple[str, IO]], list[tuple[str, IO]], int, bool, dict[str, int]]:
         Bel internal ports, bel external ports, bel config ports, bel shared ports, number of configuration bit in the bel,
         whether the bel have UserCLK, and the bel config bit mapping.
     """
-    internal: List[Tuple[str, IO]] = []
-    external: List[Tuple[str, IO]] = []
-    config: List[Tuple[str, IO]] = []
-    shared: List[Tuple[str, IO]] = []
+    internal: list[tuple[str, IO]] = []
+    external: list[tuple[str, IO]] = []
+    config: list[tuple[str, IO]] = []
+    shared: list[tuple[str, IO]] = []
     isExternal = False
     isConfig = False
     isShared = False
@@ -727,14 +626,14 @@ def parseFileVHDL(filename: str, belPrefix: str = "") -> Tuple[
     return internal, external, config, shared, noConfigBits, userClk, belMapDic
 
 
-def parseFileVerilog(filename: str, belPrefix: str = "") -> Tuple[
-    List[Tuple[str, IO]],
-    List[Tuple[str, IO]],
-    List[Tuple[str, IO]],
-    List[Tuple[str, IO]],
+def parseFileVerilog(filename: str, belPrefix: str = "") -> tuple[
+    list[tuple[str, IO]],
+    list[tuple[str, IO]],
+    list[tuple[str, IO]],
+    list[tuple[str, IO]],
     int,
     bool,
-    Dict[str, Dict],
+    dict[str, dict],
 ]:
     """
     Parse a Verilog bel file and return all the related information of the bel. The tuple returned for relating to ports
@@ -800,14 +699,14 @@ def parseFileVerilog(filename: str, belPrefix: str = "") -> Tuple[
         ValueError: No permission to access the file
 
     Returns:
-        Tuple[List[Tuple[str, IO]], List[Tuple[str, IO]], List[Tuple[str, IO]], List[Tuple[str, IO]], int, bool, Dict[str, Dict]]:
+        tuple[list[tuple[str, IO]], list[tuple[str, IO]], list[tuple[str, IO]], list[tuple[str, IO]], int, bool, dict[str, dict]]:
         Bel internal ports, bel external ports, bel config ports, bel shared ports, number of configuration bit in the bel,
         whether the bel have UserCLK, and the bel config bit mapping.
     """
-    internal: List[Tuple[str, IO]] = []
-    external: List[Tuple[str, IO]] = []
-    config: List[Tuple[str, IO]] = []
-    shared: List[Tuple[str, IO]] = []
+    internal: list[tuple[str, IO]] = []
+    external: list[tuple[str, IO]] = []
+    config: list[tuple[str, IO]] = []
+    shared: list[tuple[str, IO]] = []
     isExternal = False
     isConfig = False
     isShared = False
@@ -880,7 +779,7 @@ def parseFileVerilog(filename: str, belPrefix: str = "") -> Tuple[
 
 def _belMapProcessing(
     file: str, filename: str, syntax: Literal["vhdl", "verilog"]
-) -> Dict:
+) -> dict:
     pre = ""
     if syntax == "vhdl":
         pre = "--.*?"
@@ -954,50 +853,9 @@ def _belMapProcessing(
     return belMapDic
 
 
-def parseMatrix(fileName: str, tileName: str) -> Dict[str, List[str]]:
-    """
-    parse the matrix csv into a dictionary from destination to source
-
-    Args:
-        fileName (str): directory of the matrix csv file
-        tileName (str): name of the tile need to be parsed
-
-    Raises:
-        ValueError: Non matching matrix file content and tile name
-
-    Returns:
-        Dict[str, List[str]]: dictionary from destination to a list of source
-    """
-
-    connectionsDic = {}
-    with open(fileName, "r") as f:
-        file = f.read()
-        file = re.sub(r"#.*", "", file)
-        file = file.split("\n")
-
-    if file[0].split(",")[0] != tileName:
-        print(fileName)
-        print(file[0].split(","))
-        print(tileName)
-        raise ValueError(
-            "Tile name (top left element) in csv file does not match tile name in tile object"
-        )
-
-    destList = file[0].split(",")[1:]
-
-    for i in file[1:]:
-        i = i.split(",")
-        portName, connections = i[0], i[1:]
-        if portName == "":
-            continue
-        indices = [k for k, v in enumerate(connections) if v == "1"]
-        connectionsDic[portName] = [destList[j] for j in indices]
-    return connectionsDic
-
-
 def parseConfigMem(
     fileName: str, maxFramePerCol: int, frameBitPerRow: int, globalConfigBits: int
-) -> List[ConfigMem]:
+) -> list[ConfigMem]:
     """
     Parse the config memory csv file into a list of ConfigMem objects
 
@@ -1017,7 +875,7 @@ def parseConfigMem(
         ValueError: Invalid range entry in config bit range
 
     Returns:
-        List[ConfigMem]: _description_
+        list[ConfigMem]: _description_
     """
     with open(fileName) as f:
         mappingFile = list(csv.DictReader(f))
