+++ FABulous/geometry_generator/wire_geometry.py
@@ -1,7 +1,7 @@
 from csv import writer as csvWriter
 from typing import List
 
-from FABulous.fabric_generator.fabric import Direction
+from FABulous.fabric_definition.Fabric import Direction
 from FABulous.geometry_generator.geometry_obj import Location
